requirements.txt
aiohttp[speedups]
requests
sqlalchemy
asyncpg
pydantic
pytest
psycopg2-binary
python-dotenv
bcrypt